Only add entry pairs that are in proper temporal order
E.g. if event-stream-processor fails to log start but logs finish,
and bulker logs start but fails to log successful finish (because
event-stream-processor already did it), and event-stream-processor
finishes before bulker tries to start, the current matchers will
make it appear that the task process finished before it started,
contributing to negative times in the histogram.